requirements.txt
asgiref==3.11.0
attrs==25.4.0
charset-normalizer==3.4.4
colorama==0.4.6
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
reportlab==4.4.6
rpds-py==0.30.0
sqlparse==0.5.4
tzdata==2025.3
uritemplate==4.2.0
whitenoise==6.11.0
